N64: Better analog nub acceleration, and added -mhard-float in CXXFLAGS when building
svn-id: r47708

+++ backends/platform/n64/osys_n64.h
@@ -115,7 +115,7 @@ protected:
 
 	bool _mouseVisible;
 	volatile int _mouseX, _mouseY;
-	volatile int _tempMouseX, _tempMouseY;
+	volatile float _tempMouseX, _tempMouseY;
 	volatile int _mouseMaxX, _mouseMaxY;
 	int _mouseHotspotX, _mouseHotspotY;
